requirements.txt
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
selenium==4.28.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
